Remove synchronous cache lookup from remote execution.
[ci skip-build-wheels]
