+++ src/rust/engine/process_execution/src/remote_tests.rs
@@ -21,7 +21,7 @@ use testutil::data::{TestData, TestDirectory, TestTree};
 use testutil::{owned_string_vec, relative_paths};
 use workunit_store::{RunId, WorkunitStore};
 
-use crate::remote::{digest, CommandRunner, ExecutionError, OperationOrStatus};
+use crate::remote::{CommandRunner, ExecutionError, OperationOrStatus};
 use crate::{
   CommandRunner as CommandRunnerTrait, Context, FallibleProcessResultWithPlatform, InputDigests,
   Platform, Process, ProcessCacheScope, ProcessError, ProcessMetadata,
@@ -36,8 +36,6 @@ const RETRY_INTERVAL: Duration = Duration::from_micros(0);
 const STORE_CONCURRENCY_LIMIT: usize = 256;
 const STORE_BATCH_API_SIZE_LIMIT: usize = 4 * 1024 * 1024;
 const EXEC_CONCURRENCY_LIMIT: usize = 256;
-const CACHE_CONCURRENCY_LIMIT: usize = 256;
-const CACHE_READ_TIMEOUT: Duration = Duration::from_secs(5);
 
 #[derive(Debug, PartialEq)]
 pub(crate) struct RemoteTestResult {
@@ -662,32 +660,25 @@ async fn successful_with_only_call_to_execute() {
   let op_name = "gimme-foo".to_string();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
-    let action_digest = digest(&action).unwrap();
 
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("")),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Ok(vec![
-            make_incomplete_operation(&op_name),
-            make_successful_operation(
-              &op_name,
-              StdoutType::Raw("foo".to_owned()),
-              StderrType::Raw("".to_owned()),
-              0,
-            ),
-          ]),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Ok(vec![
+          make_incomplete_operation(&op_name),
+          make_successful_operation(
+            &op_name,
+            StdoutType::Raw("foo".to_owned()),
+            StderrType::Raw("".to_owned()),
+            0,
+          ),
+        ]),
+      }]),
       None,
     )
   };
@@ -710,19 +701,14 @@ async fn successful_after_reconnect_with_wait_execution() {
   let op_name = "gimme-foo".to_string();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
-    let action_digest = digest(&action).unwrap();
 
     mock::execution_server::TestServer::new(
       mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
         ExpectedAPICall::Execute {
           execute_request,
           stream_responses: Ok(vec![make_incomplete_operation(&op_name)]),
@@ -760,7 +746,7 @@ async fn successful_after_reconnect_from_retryable_error() {
   let op_name_2 = "gimme-bar".to_string();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
@@ -768,14 +754,8 @@ async fn successful_after_reconnect_from_retryable_error() {
 
     let execute_request_2 = execute_request.clone();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
       mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
         ExpectedAPICall::Execute {
           execute_request,
           stream_responses: Ok(vec![
@@ -812,74 +792,21 @@ async fn successful_after_reconnect_from_retryable_error() {
 }
 
 #[tokio::test]
-async fn successful_served_from_action_cache() {
-  WorkunitStore::setup_for_tests();
-  let execute_request = echo_foo_request();
-
-  let mock_server = {
-    let (action, _, _) = crate::remote::make_execute_request(
-      &execute_request.clone().try_into().unwrap(),
-      ProcessMetadata::default(),
-    )
-    .unwrap();
-
-    let action_digest = digest(&action).unwrap();
-
-    let action_result = make_action_result(
-      StdoutType::Raw("foo".to_owned()),
-      StderrType::Raw("".to_owned()),
-      0,
-      None,
-    );
-
-    mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::GetActionResult {
-        action_digest,
-        response: Ok(action_result),
-      }]),
-      None,
-    )
-  };
-
-  let result = run_command_remote(mock_server.address(), execute_request)
-    .await
-    .unwrap();
-
-  assert_eq!(result.stdout_bytes, "foo".as_bytes());
-  assert_eq!(result.stderr_bytes, "".as_bytes());
-  assert_eq!(result.original.exit_code, 0);
-  assert_eq!(result.original.output_directory, *EMPTY_DIRECTORY_DIGEST);
-  assert_cancellation_requests(&mock_server, vec![]);
-}
-
-#[tokio::test]
 async fn server_rejecting_execute_request_gives_error() {
   WorkunitStore::setup_for_tests();
 
   let execute_request = echo_foo_request();
 
   let mock_server = mock::execution_server::TestServer::new(
-    mock::execution_server::MockExecution::new(vec![
-      ExpectedAPICall::GetActionResult {
-        action_digest: hashing::Digest::new(
-          hashing::Fingerprint::from_hex_string(
-            "bf10cd168ad711602f7a241cbcbc9a3d32497fdd1465d8a01d4549ee7d8ebc08",
-          )
-          .unwrap(),
-          142,
-        ),
-        response: Err(Status::not_found("")),
-      },
-      ExpectedAPICall::Execute {
-        execute_request: crate::remote::make_execute_request(
-          &Process::new(owned_string_vec(&["/bin/echo", "-n", "bar"])),
-          ProcessMetadata::default(),
-        )
-        .unwrap()
-        .2,
-        stream_responses: Err(Status::invalid_argument("".to_owned())),
-      },
-    ]),
+    mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+      execute_request: crate::remote::make_execute_request(
+        &Process::new(owned_string_vec(&["/bin/echo", "-n", "bar"])),
+        ProcessMetadata::default(),
+      )
+      .unwrap()
+      .2,
+      stream_responses: Err(Status::invalid_argument("".to_owned())),
+    }]),
     None,
   );
 
@@ -897,25 +824,17 @@ async fn server_sending_triggering_timeout_with_deadline_exceeded() {
   let execute_request = echo_foo_request();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Err(Status::deadline_exceeded("")),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Err(Status::deadline_exceeded("")),
+      }]),
       None,
     )
   };
@@ -934,33 +853,25 @@ async fn sends_headers() {
   let op_name = "gimme-foo".to_string();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Ok(vec![
-            make_incomplete_operation(&op_name),
-            make_successful_operation(
-              &op_name,
-              StdoutType::Raw("foo".to_owned()),
-              StderrType::Raw("".to_owned()),
-              0,
-            ),
-          ]),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Ok(vec![
+          make_incomplete_operation(&op_name),
+          make_successful_operation(
+            &op_name,
+            StdoutType::Raw("foo".to_owned()),
+            StderrType::Raw("".to_owned()),
+            0,
+          ),
+        ]),
+      }]),
       None,
     )
   };
@@ -985,7 +896,6 @@ async fn sends_headers() {
 
   let command_runner = CommandRunner::new(
     &mock_server.address(),
-    &mock_server.address(),
     ProcessMetadata::default(),
     None,
     btreemap! {
@@ -997,8 +907,6 @@ async fn sends_headers() {
     OVERALL_DEADLINE_SECS,
     RETRY_INTERVAL,
     EXEC_CONCURRENCY_LIMIT,
-    CACHE_CONCURRENCY_LIMIT,
-    CACHE_READ_TIMEOUT,
     None,
   )
   .unwrap();
@@ -1017,7 +925,7 @@ async fn sends_headers() {
     .iter()
     .map(|received_message| received_message.headers.clone())
     .collect();
-  assert_that!(message_headers).has_length(2);
+  assert_that!(message_headers).has_length(1);
   for headers in message_headers {
     {
       let want_key = "google.devtools.remoteexecution.v1test.requestmetadata-bin";
@@ -1137,33 +1045,25 @@ async fn ensure_inline_stdio_is_stored() {
   let mock_server = {
     let op_name = "cat".to_owned();
 
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &echo_roland_request().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Ok(vec![
-            make_incomplete_operation(&op_name),
-            make_successful_operation(
-              &op_name.clone(),
-              StdoutType::Raw(test_stdout.string()),
-              StderrType::Raw(test_stderr.string()),
-              0,
-            ),
-          ]),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Ok(vec![
+          make_incomplete_operation(&op_name),
+          make_successful_operation(
+            &op_name.clone(),
+            StdoutType::Raw(test_stdout.string()),
+            StderrType::Raw(test_stderr.string()),
+            0,
+          ),
+        ]),
+      }]),
       None,
     )
   };
@@ -1190,7 +1090,6 @@ async fn ensure_inline_stdio_is_stored() {
 
   let cmd_runner = CommandRunner::new(
     &mock_server.address(),
-    &mock_server.address(),
     ProcessMetadata::default(),
     None,
     BTreeMap::new(),
@@ -1199,8 +1098,6 @@ async fn ensure_inline_stdio_is_stored() {
     OVERALL_DEADLINE_SECS,
     RETRY_INTERVAL,
     EXEC_CONCURRENCY_LIMIT,
-    CACHE_CONCURRENCY_LIMIT,
-    CACHE_READ_TIMEOUT,
     None,
   )
   .unwrap();
@@ -1283,38 +1180,30 @@ async fn initial_response_error() {
   let mock_server = {
     let op_name = "gimme-foo".to_string();
 
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Ok(vec![MockOperation::new({
-            Operation {
-              name: op_name.to_string(),
-              done: true,
-              result: Some(protos::gen::google::longrunning::operation::Result::Error(
-                protos::gen::google::rpc::Status {
-                  code: Code::Internal as i32,
-                  message: "Something went wrong".to_string(),
-                  ..Default::default()
-                },
-              )),
-              ..Default::default()
-            }
-          })]),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Ok(vec![MockOperation::new({
+          Operation {
+            name: op_name.to_string(),
+            done: true,
+            result: Some(protos::gen::google::longrunning::operation::Result::Error(
+              protos::gen::google::rpc::Status {
+                code: Code::Internal as i32,
+                message: "Something went wrong".to_string(),
+                ..Default::default()
+              },
+            )),
+            ..Default::default()
+          }
+        })]),
+      }]),
       None,
     )
   };
@@ -1333,31 +1222,23 @@ async fn initial_response_missing_response_and_error() {
   let mock_server = {
     let op_name = "gimme-foo".to_string();
 
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
-        ExpectedAPICall::Execute {
-          execute_request,
-          stream_responses: Ok(vec![MockOperation::new({
-            Operation {
-              name: op_name.to_string(),
-              done: true,
-              ..Default::default()
-            }
-          })]),
-        },
-      ]),
+      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::Execute {
+        execute_request,
+        stream_responses: Ok(vec![MockOperation::new({
+          Operation {
+            name: op_name.to_string(),
+            done: true,
+            ..Default::default()
+          }
+        })]),
+      }]),
       None,
     )
   };
@@ -1378,20 +1259,14 @@ async fn fails_after_retry_limit_exceeded() {
   let execute_request = echo_foo_request();
 
   let mock_server = {
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
       mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
         ExpectedAPICall::Execute {
           execute_request: execute_request.clone(),
           stream_responses: Ok(vec![make_retryable_operation_failure()]),
@@ -1438,20 +1313,14 @@ async fn fails_after_retry_limit_exceeded_with_stream_close() {
 
   let mock_server = {
     let op_name = "foo-bar".to_owned();
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &execute_request.clone().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
       mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
         ExpectedAPICall::Execute {
           execute_request: execute_request.clone(),
           stream_responses: Ok(vec![make_incomplete_operation(&op_name)]),
@@ -1501,20 +1370,14 @@ async fn execute_missing_file_uploads_if_known() {
   let mock_server = {
     let op_name = "cat".to_owned();
 
-    let (action, _, execute_request) = crate::remote::make_execute_request(
+    let (_, _, execute_request) = crate::remote::make_execute_request(
       &cat_roland_request().try_into().unwrap(),
       ProcessMetadata::default(),
     )
     .unwrap();
 
-    let action_digest = digest(&action).unwrap();
-
     mock::execution_server::TestServer::new(
       mock::execution_server::MockExecution::new(vec![
-        ExpectedAPICall::GetActionResult {
-          action_digest,
-          response: Err(Status::not_found("".to_owned())),
-        },
         ExpectedAPICall::Execute {
           execute_request,
           stream_responses: Ok(vec![
@@ -1575,7 +1438,6 @@ async fn execute_missing_file_uploads_if_known() {
     .expect("Saving directory bytes to store");
   let command_runner = CommandRunner::new(
     &mock_server.address(),
-    &mock_server.address(),
     ProcessMetadata::default(),
     None,
     BTreeMap::new(),
@@ -1584,8 +1446,6 @@ async fn execute_missing_file_uploads_if_known() {
     OVERALL_DEADLINE_SECS,
     RETRY_INTERVAL,
     EXEC_CONCURRENCY_LIMIT,
-    CACHE_CONCURRENCY_LIMIT,
-    CACHE_READ_TIMEOUT,
     None,
   )
   .unwrap();
@@ -1612,16 +1472,7 @@ async fn execute_missing_file_errors_if_unknown() {
 
   let mock_server = {
     mock::execution_server::TestServer::new(
-      mock::execution_server::MockExecution::new(vec![ExpectedAPICall::GetActionResult {
-        action_digest: hashing::Digest::new(
-          hashing::Fingerprint::from_hex_string(
-            "3174d44cc8d68ec5047f2da24a4db150ab0264a9d6558872081585696a98b410",
-          )
-          .unwrap(),
-          144,
-        ),
-        response: Err(Status::not_found("".to_owned())),
-      }]),
+      mock::execution_server::MockExecution::new(vec![]),
       None,
     )
   };
@@ -1647,7 +1498,6 @@ async fn execute_missing_file_errors_if_unknown() {
 
   let runner = CommandRunner::new(
     &mock_server.address(),
-    &mock_server.address(),
     ProcessMetadata::default(),
     None,
     BTreeMap::new(),
@@ -1656,8 +1506,6 @@ async fn execute_missing_file_errors_if_unknown() {
     OVERALL_DEADLINE_SECS,
     RETRY_INTERVAL,
     EXEC_CONCURRENCY_LIMIT,
-    CACHE_CONCURRENCY_LIMIT,
-    CACHE_READ_TIMEOUT,
     None,
   )
   .unwrap();
@@ -2322,7 +2170,6 @@ fn create_command_runner(
   let store = make_store(store_dir.path(), cas, runtime.clone());
   let command_runner = CommandRunner::new(
     &address,
-    &address,
     ProcessMetadata::default(),
     None,
     BTreeMap::new(),
@@ -2331,8 +2178,6 @@ fn create_command_runner(
     OVERALL_DEADLINE_SECS,
     RETRY_INTERVAL,
     EXEC_CONCURRENCY_LIMIT,
-    CACHE_CONCURRENCY_LIMIT,
-    CACHE_READ_TIMEOUT,
     None,
   )
   .expect("Failed to make command runner");
